将oss敏感字段写在了.env文件中,同时完成了自动填写公共字段以及添加菜品、文件上传相关功能，引入了.env文件，并将其写在.gitignore，不提交

+++ sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -35,7 +35,7 @@ public class CategoryController {
     }
 
     @DeleteMapping()
-    public Result deleteById(int id){
+    public Result deleteById(long id){
         log.info("需要删除的分类id为：{}", id);
         categoryService.deleteById(id);
         return Result.success();
@@ -49,7 +49,7 @@ public class CategoryController {
     }
 
     @PostMapping("/status/{status}")
-    public Result changeStatus(@PathVariable int status, int id){
+    public Result changeStatus(@PathVariable int status, long id){
         log.info("更改的id为{}，状态为{}",id,status);
         categoryService.changeStatus(status, id);
         return Result.success();
